+++ src/hittable.rs
@@ -1,12 +1,14 @@
+use std::sync::Arc;
 use crate::common::*;
 use crate::ray::Ray;
 pub use crate::interval::Interval;
+use crate::material::Material;
 
-#[derive(Clone, Copy, Debug)]
 pub struct HitRecord {
     pub point: Point3,
     pub normal: Vec3,
     pub t: f64,
+    pub material: Arc<dyn Material>,
 }
 
 pub trait Hittable {