+++ src/main.rs
@@ -20,19 +20,29 @@ fn main() {
     // World
     let mut world = HittableList::new();
 
+    let meterial_ground: Arc<dyn Material> = Arc::new(material::Lambertian::new(Color::new(0.5, 0.5, 0.8)));
     let material_center: Arc<dyn Material> = Arc::new(material::Lambertian::new(Color::new(0.1, 0.2, 0.5)));
+    let material_left: Arc<dyn Material> = Arc::new(material::Metal::new(Color::new(0.8, 0.8, 0.8)));
+    let material_right: Arc<dyn Material> = Arc::new(material::Metal::new(Color::new(0.8, 0.6, 0.2)));
 
     world.add(Box::new(
-        Sphere::new(Point3::new(0.0, 0.0, -1.0), 0.5, Arc::clone(&material_center))
+        Sphere::new(Point3::new(0.0, -100.5, -1.0), 100.0, Arc::clone(&meterial_ground))
     ));
     world.add(Box::new(
-        Sphere::new(Point3::new(0.0, -100.5, -1.0), 100.0, Arc::clone(&material_center))
+        Sphere::new(Point3::new(0.0, 0.0, -1.2), 0.5, Arc::clone(&material_center))
     ));
+    world.add(Box::new(
+        Sphere::new(Point3::new(-1.0, 0.0, -1.0), 0.5, Arc::clone(&material_left))
+    ));
+    world.add(Box::new(
+        Sphere::new(Point3::new(1.0, 0.0, -1.0), 0.5, Arc::clone(&material_right))
+    ));
+
 
     // Camera
     let camera = Camera::new(16.0 / 9.0, 1920);
     let (width, height, buffer) = camera.render(&world);
-    let file_name = "output/material.jpg";
+    let file_name = "output/material_metal.jpg";
     match write_jpg(file_name, width, height, &buffer, 100) {
         Ok(()) => println!("Wrote {} ({}x{} pixels)", file_name, width, height),
         Err(e) => eprintln!("Failed to write {}: {}", file_name, e),